requirements.txt
apache-airflow[crypto,s3,password,postgres]==1.9.0
boto3==1.7.4
psycopg2==2.7.5
requests==2.19.0
werkzeug<1.0